fix: rename Pannel => Panel
