+++ src/components/Panel/index.js
